feat: massive refactor
