+++ src/services/finance_analyzer/constants.py
